pyproject.toml
[project]
name = "sparkle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0b251028",
    "agent-framework-azure-ai==1.0.0b251028",
    "azure-ai-agents==1.2.0b5",
    "azure-identity>=1.17.0",
]
